+++ packages/radonis-query/src/index.ts
@@ -9,3 +9,4 @@
 
 export { useQuery } from './hooks/useQuery'
 export { queryPlugin } from './plugin'
+export { createQueryPrefetcher } from './queryPrefetcher'