+++ packages/radonis-server/src/React/components/HydrationRoot.tsx
@@ -16,7 +16,7 @@ import { useAssetsManager } from '../hooks/internal/useAssetsManager'
 import { useManifestBuilder } from '../hooks/internal/useManifestBuilder'
 
 /**
- * Component for drawing the line between parts of the page
+ * The component for drawing the line between parts of the page
  * that should and should not be hydrated client-side
  * @see {@link https://radonis.vercel.app/docs/components#hydrating-components}
  */